Ensured spawn message order

+++ GoSentinel/Models/PokemonSpawn.cs
@@ -9,8 +9,8 @@ namespace GoSentinel.Models
     {
         public string SpawnpointId { get; set; }
         public int PokemonId { get; set; }
-        public double Latitude { get; set; }
-        public double Longitude { get; set; }
+        public float Latitude { get; set; }
+        public float Longitude { get; set; }
         public DateTime DisappearTime { get; set; }
         public int Attack { get; set; }
         public int Defense { get; set; }
